requirements.txt
requests==2.28.1
urllib3==1.26.13
xlrd3==1.1.0
PyYAML==6.0
PyMySQL==1.0.2
DBUtils==3.0.2
redis==4.4.2
jsoncomparison==1.1.0
pytest==7.2.0
dbutils==3.0.2
allure-pytest==2.12.0
allure-python-commons==2.12.0